+++ dsr_agents/src/agents/nav_agent.cpp
@@ -22,7 +22,7 @@
 
 // DSR
 #include "dsr_agents/qt_executor.hpp"
-#include "dsr_agents/nav_agent.hpp"
+#include "dsr_agents/agents/nav_agent.hpp"
 
 /* Initialize the publishers and subscribers */
 navigationAgent::navigationAgent(): AgentNode("navigation_agent"), current_room_(""){