+++ dsr_agents/src/agents/tf_agent.cpp
@@ -14,7 +14,7 @@
 
 // DSR
 #include "dsr_agents/qt_executor.hpp"
-#include "dsr_agents/tf_agent.hpp"
+#include "dsr_agents/agents/tf_agent.hpp"
 
 /* Initialize the publishers and subscribers */
 tfAgent::tfAgent(): AgentNode("tf_agent"){